fix(main,shared): store exchange rates as scaled integers, not floats
Exchange rate and transfer.exchange_rate were stored as raw decimals in
SQLite, even though domain/ExchangeRate.ts already had fromStored/toStored
scaling logic for this (mirroring how Money stores cents). That scaling
was dead code before this change. Now the migration, repositories, and
services actually use it, and formatted fields are consistently snake_case
(amount_formatted, rate_formatted, exchange_rate_formatted).

+++ packages/main/src/db/migrations/0001_initial_schema.ts
@@ -22,7 +22,7 @@ export function up(db: Database.Database): void {
       id            INTEGER PRIMARY KEY AUTOINCREMENT,
       from_currency TEXT    NOT NULL CHECK (from_currency IN ('TRY', 'USD', 'CAD', 'EUR')),
       to_currency   TEXT    NOT NULL CHECK (to_currency   IN ('TRY', 'USD', 'CAD', 'EUR')),
-      rate          REAL    NOT NULL,
+      rate          INTEGER NOT NULL,
       recorded_at   TEXT    NOT NULL,
       CHECK (from_currency != to_currency)
     );
@@ -42,7 +42,7 @@ export function up(db: Database.Database): void {
       id                  INTEGER PRIMARY KEY AUTOINCREMENT,
       from_transaction_id INTEGER NOT NULL REFERENCES transactions(id),
       to_transaction_id   INTEGER NOT NULL REFERENCES transactions(id),
-      exchange_rate       REAL,
+      exchange_rate       INTEGER,
       CHECK (from_transaction_id != to_transaction_id)
     );
   `)
